Massive refactoring, part 1.
Fixture is responsible for only one device (also host processor device is possible in new architecture), simplified fixture runner (contexts and queues are created in separate class responsible for managing a particular device), benchmark results are stored in JSON file, HTML builder will be removed (it's hard to maintain and extend, will replace with a spreadsheet that would be based on data serialized in JSON). All fixtures except trivial factorial are disabled for now, all indicators except duration are not implemented yet.